+++ src/Samples/NewWindow/Program.fs
@@ -2,7 +2,8 @@
 
 open System
 open System.Windows
-open Elmish
+open Serilog
+open Serilog.Extensions.Logging
 open Elmish.WPF
 
 module App =
@@ -119,14 +120,21 @@ let window2DesignVm = ViewModel.designInstance App.initWindow2 (App.window2Bindi
 
 
 let main mainWindow (createWindow1: Func<#Window>) (createWindow2: Func<#Window>) =
+
+  let logger =
+    LoggerConfiguration()
+      .MinimumLevel.Override("Elmish.WPF.Update", Events.LogEventLevel.Verbose)
+      .MinimumLevel.Override("Elmish.WPF.Bindings", Events.LogEventLevel.Verbose)
+      .MinimumLevel.Override("Elmish.WPF.Performance", Events.LogEventLevel.Verbose)
+      .WriteTo.Console()
+      .CreateLogger()
+
   let createWindow1 () = createWindow1.Invoke()
   let createWindow2 () =
     let window = createWindow2.Invoke()
     window.Owner <- mainWindow
     window
   let bindings = App.mainBindings createWindow1 createWindow2
-  Program.mkSimpleWpf App.init App.update bindings
-  |> Program.withConsoleTrace
-  |> Program.runWindowWithConfig
-    { ElmConfig.Default with LogConsole = true; Measure = true }
-    mainWindow
+  WpfProgram.mkSimple App.init App.update bindings
+  |> WpfProgram.withLogger (new SerilogLoggerFactory(logger))
+  |> WpfProgram.runWindow mainWindow