+++ src/Samples/SubModelOpt/Program.fs
@@ -1,6 +1,7 @@
 ﻿module Elmish.WPF.Samples.SubModelOpt.Program
 
-open Elmish
+open Serilog
+open Serilog.Extensions.Logging
 open Elmish.WPF
 
 
@@ -121,8 +122,15 @@ let mainDesignVm = ViewModel.designInstance (App.init ()) (App.bindings ())
 
 
 let main window =
-  Program.mkSimpleWpf App.init App.update App.bindings
-  |> Program.withConsoleTrace
-  |> Program.runWindowWithConfig
-    { ElmConfig.Default with LogConsole = true; Measure = true }
-    window
+
+  let logger =
+    LoggerConfiguration()
+      .MinimumLevel.Override("Elmish.WPF.Update", Events.LogEventLevel.Verbose)
+      .MinimumLevel.Override("Elmish.WPF.Bindings", Events.LogEventLevel.Verbose)
+      .MinimumLevel.Override("Elmish.WPF.Performance", Events.LogEventLevel.Verbose)
+      .WriteTo.Console()
+      .CreateLogger()
+
+  WpfProgram.mkSimple App.init App.update App.bindings
+  |> WpfProgram.withLogger (new SerilogLoggerFactory(logger))
+  |> WpfProgram.runWindow window